+++ src/constants.ts
@@ -78,6 +78,7 @@ export enum ClientSubredditJob {
     SendDailyDigest = "sendDailyDigest",
     NotifyModTeamOnDemod = "notifyModTeamOnDemod",
     PermissionCheckEnqueue = "permissionCheckEnqueue",
+    ClientSubAnnouncements = "clientSubAnnouncements",
 }
 
 /**