+++ src/handleClientPostOrComment.ts
@@ -19,6 +19,10 @@ export async function handleClientPostCreate (event: PostCreate, context: Trigge
         throw new Error("Content Create: handleClientPostCreate should not be called for the control subreddit, check the subreddit name handling logic");
     }
 
+    if (await context.settings.get<boolean>(AppSetting.DisableClientChecks)) {
+        return;
+    }
+
     event = await fixPostTriggerEvent(event, context);
 
     if (!event.post || !event.author?.name) {
@@ -64,6 +68,10 @@ export async function handleClientPostCreate (event: PostCreate, context: Trigge
 export async function handleClientCommentCreate (event: CommentCreate, context: TriggerContext) {
     if (context.subredditName === CONTROL_SUBREDDIT) {
         throw new Error("Content Create: handleClientCommentCreate should not be called for the control subreddit, check the subreddit name handling logic");
+    }
+
+    if (await context.settings.get<boolean>(AppSetting.DisableClientChecks)) {
+        return;
     }
 
     const fixedEvent = await fixCommentTriggerEvent(event, context);
@@ -123,6 +131,10 @@ export async function handleClientCommentUpdate (event: CommentUpdate, context: 
         return;
     }
 
+    if (await context.settings.get<boolean>(AppSetting.DisableClientChecks)) {
+        return;
+    }
+
     const fixedEvent = await fixCommentTriggerEvent(event, context);
 
     if (!fixedEvent.comment || !fixedEvent.author?.name) {
@@ -178,6 +190,10 @@ export async function handleClientCommentUpdate (event: CommentUpdate, context: 
 
 export async function handleClientPostUpdate (event: PostUpdate, context: TriggerContext) {
     if (context.subredditName === CONTROL_SUBREDDIT) {
+        return;
+    }
+
+    if (await context.settings.get<boolean>(AppSetting.DisableClientChecks)) {
         return;
     }
 
@@ -277,7 +293,7 @@ async function handleContentCreation (username: string, currentStatus: UserDetai
         const isCurrentlyBanned = await isBanned(context.reddit, subredditName, user.username);
 
         if (!isCurrentlyBanned) {
-            let message = await context.settings.get<string>(AppSetting.BanMessage) ?? CONFIGURATION_DEFAULTS.banMessage;
+            let message = settings[AppSetting.BanMessage] as string | undefined ?? CONFIGURATION_DEFAULTS.banMessage;
             message = message.replaceAll("{subreddit}", subredditName)
                 .replaceAll("{account}", user.username)
                 .replaceAll("{link}", user.username);